[FIX] Fix product..components.ts for the cards. Review averageRating
