+++ src/app/core/interfaces/comment.ts
@@ -1,0 +1,6 @@
+export interface Comment {
+  rating: number;
+  text: string;
+  userId: number;
+  productId: number;
+}
